Fully replace filesystem with filesystem2 on Darwin and fix Linux specs
